+++ packages/pdf_document/lib/pdf_document.dart
@@ -7,6 +7,7 @@ export 'src/content_elements.dart';
 export 'src/content_writer.dart';
 export 'src/document.dart';
 export 'src/editor.dart';
+export 'src/font_embedder.dart';
 export 'src/form.dart';
 export 'src/image.dart';
 export 'src/image_pdf.dart';